refactor(site): share confirm delete dialog
