+++ site/src/pages/AgentsPage/components/LimitsTab/GroupLimitsSection.tsx
@@ -6,6 +6,7 @@ import type { ChatUsageLimitGroupOverride, Group } from "#/api/typesGenerated";
 import { Autocomplete } from "#/components/Autocomplete/Autocomplete";
 import { AvatarData } from "#/components/Avatar/AvatarData";
 import { Button } from "#/components/Button/Button";
+import { ConfirmDeleteDialog } from "#/components/Dialogs/ConfirmDeleteDialog/ConfirmDeleteDialog";
 import { Input } from "#/components/Input/Input";
 import { Label } from "#/components/Label/Label";
 import { Spinner } from "#/components/Spinner/Spinner";
@@ -22,7 +23,6 @@ import {
 	formatCostMicros,
 	isPositiveFiniteDollarAmount,
 } from "#/utils/currency";
-import { ConfirmDeleteDialog } from "../ConfirmDeleteDialog";
 import { SectionHeader } from "../SectionHeader";
 
 interface GroupLimitsSectionProps {
